+++ common/idm/src/idm_session.c
@@ -183,7 +183,8 @@ bool IsSessionTimeout(void)
     }
     uint64_t currentTime = GetSystemTime();
     if (currentTime < g_session->time || currentTime - g_session->time > SESSION_VALIDITY_PERIOD) {
-        LOG_ERROR("timeout, %{public}llu, %{public}llu", currentTime, g_session->time);
+        LOG_ERROR("timeout, %{public}llu, %{public}llu",
+                  (unsigned long long)currentTime, (unsigned long long)g_session->time);
         return true;
     }
     return false;
